+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 2026-02-09
 - Gamble factions to indicate supporters for each gambler or outside entities within the gamble
 - Auto detection for media type during submission
+- Pagination updated to check for url
 
 ## 2026-02-08
 - Media default view updated to show general (excluding volumes)